Improved Main to work with the CC monad.
To make the whole thing work, it seems quite necessary
to limit the scope of delimited continuations to one iteration
of the REPL. This may be deep or may not be, should investigate
further.

It also seemed that it was necessary to rewrite loadIO in a
non-point-free style. This puzzles me because I though that
f x = g (h x) can always be rewritten as f = g . h; however
this transformation seems to prevent the program from being
correctly typed. This is almost certainly an interaction
between RankNTypes and point-free style. I should investigate
this further.